Create main.cpp
Week 1 read me 

Assignment 1 The program shows how to create and be able to use string attributes and file input and output operations. I was able to create a class called Item that has two public data members: name and quantity. It also includes a saveToFile() method that writes the item name and quantity into a new file that I created called 'items.text', and another method called loadFromFile() that reads the file.